WT-13388 Move general extent list utilities into utils_extlist.cpp. Also delete some unneeded comments.
