+++ test/unittest/CMakeLists.txt
@@ -34,6 +34,7 @@ endif()
 set(unittest_sources
     tests/main.cpp
     tests/utils.cpp
+    tests/utils_extlist.cpp
     tests/wrappers/block_mods.cpp
     tests/wrappers/connection_wrapper.cpp
     tests/wrappers/item_wrapper.cpp
